requirements.txt
streamlit==1.38.0
google-generativeai>=0.8.3
python-dotenv>=1.0.0
chromadb==0.5.5
pandas>=2.0.0
pypdf2>=3.0.0
tabulate>=0.9.0
protobuf==5.29.3
rich==13.9.4
tenacity==8.5.0
numpy<2.0.0